backend/requirements.txt
Flask
flask-cors
sentence-transformers
faiss-cpu
pandas
numpy
ollama